move get_latest_versions to common utils
